requirements.txt
httpx>=0.27
beautifulsoup4>=4.12
lxml>=5.0
markdownify>=0.12
pyyaml>=6.0
typer>=0.12
rich>=13.0
playwright>=1.44
